Added more details to History (epoch rocs), and tuning every certain number of epochs.

+++ batching.py
@@ -140,6 +140,6 @@ def num_iters(db, collection, batch_size=None, subset_size=None):
     if not batch_size:
         batch_size = PREFERRED_BATCH_SIZES[db][collection]
     if subset_size:
-        return np.ceil(subset_size / batch_size)
+        return int(np.ceil(subset_size / batch_size))
     collection_size = stats.COLLECTION_SIZE[db][collection]
-    return np.ceil(collection_size / batch_size)
+    return int(np.ceil(collection_size / batch_size))
